Callback function all the things
Textual 6 will use callback functions for the “app” object which means
Equinox must be updated to accommodate this change.

This commit changes the following:
• NickColorGenerator.prototype.generateColorFromNickname() now takes a
callback function as a parameter. This callback function is invoked
once the one for app.nicknameColorStyleHash() has completed.
• An anonymous function is used in NickColorGenerator to maintain a
reference to inlineNicks until generateColorFromNickname() has finished.
• Removed app.channelIsJoined() check form Textual.handleEvent(). This
check is unnecessary because ‘channelJoined’ is already being checked
for.
• Updated all other code to use callback functions.